requirements.txt
streamlit
pdfplumber
python-dotenv

langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma

chromadb
huggingface-hub

google-generativeai
